requirements.txt
pandas
python-dotenv
weaviate-client
requests


